fix: utf8 path panic, tui terminal restore, depth limit, module split
Fixes truncate_path in export.rs and views.rs to use char based slicing
instead of byte indexing. Paths with non ASCII characters (Chinese,
Japanese, accented) no longer panic when truncated for display.

TUI now uses a TerminalGuard that restores raw mode and alternate screen
on drop, even if setup fails partway or the event loop errors out.

Restores the max_depth constraint on source file mtime sampling. The
breadth first walk is now bounded to 4 levels deep matching the original
design from projects.rs to avoid deep traversal into monorepos.

Splits categories.rs (was 595 lines) by extracting dev artifact
classification and cache path matching into dev_artifacts.rs (151 lines).
Production code in categories.rs is now 290 lines, under the 300 limit.
