+++ src/audit/mod.rs
@@ -5,6 +5,7 @@
 //! or saved to SQLite for historical comparison.
 
 pub mod categories;
+pub mod dev_artifacts;
 pub mod export;
 
 use std::collections::HashMap;
